# docs/openapi_admin.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "管理员端接口",
    "description": "管理员端的相关api。本地开发环境需要删除/api前缀再进行请求，服务器生产环境配置了nginx路由，直接请求即可。最终版。",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/admin/auth/login": {
      "post": {
        "summary": "管理员登录",
        "deprecated": false,
        "description": "输入用户名和密码进行登录，获取 Token。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "登录提交表单",
                "type": "object",
                "properties": {
                  "account": {
                    "title": "用户名",
                    "type": "string",
                    "description": "管理员账号",
                    "example": "admin"
                  },
                  "password": {
                    "title": "密码",
                    "type": "string",
                    "description": "管理员密码",
                    "example": "123456"
                  }
                },
                "required": [
                  "account",
                  "password"
                ]
              },
              "example": {
                "account": "admin1@test.com",
                "password": "123456"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isNewUser": {
                          "type": "boolean"
                        },
                        "hasPassword": {
                          "type": "boolean"
                        },
                        "userInfo": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "string"
                            },
                            "nickname": {
                              "type": "string"
                            },
                            "avatar": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "userId",
                            "nickname",
                            "avatar",
                            "email"
                          ]
                        },
                        "userSig": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "isNewUser",
                        "hasPassword",
                        "userInfo",
                        "userSig",
                        "token",
                        "refreshToken"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 200,
                  "message": "操作成功",
                  "data": {
                    "isNewUser": false,
                    "hasPassword": true,
                    "userInfo": {
                      "userId": "15",
                      "nickname": "elaine",
                      "avatar": "https://afternoon3-rednote.oss-cn-shenzhen.aliyuncs.com/default_avatar.jpg",
                      "email": "garhlz257@gmail.com"
                    },
                    "userSig": "eJyrVgrxCdZLrSjILEpVsjIzMLEwMNABi5WlFilZKRnpGShB*MUp2YkFBZkpSlaGZgYGhobmRibGEJnMlNS8ksy0TLAGQ1OY*sx0IDfcPSTIJTA13T8lvCzMzz0gvbzEVF87wzIzpMI-qSw4M8TYL00-yT84uyDQFqqxJDMX6BRDczMzEwtTI1PjWgArDS*b",
                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE1LCJyb2xlIjoiQURNSU4iLCJuaWNrbmFtZSI6ImVsYWluZSIsInR5cGUiOiJhY2Nlc3MiLCJleHAiOjE3NjcwOTAwNTN9.n5RqcnxwNcP1kF9NJxbqJnapxA8Fxp1RxWa_68Snsp8",
                    "refreshToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE1LCJ0eXBlIjoicmVmcmVzaCIsImV4cCI6MTc2OTA3NzI1M30.0b5cYEPz0m9BtvvgakmlI145k41w01VGZ3BI4KFFNvE"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "登录失败响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 400
                    },
                    "msg": {
                      "title": "错误提示",
                      "type": "string",
                      "example": "用户名或密码错误"
                    },
                    "data": {
                      "title": "数据",
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 400,
                  "msg": "用户名或密码错误",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/send-code": {
      "post": {
        "summary": "发送验证码",
        "deprecated": false,
        "description": "用于忘记密码或修改密码场景。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "发送验证码参数",
                "type": "object",
                "properties": {
                  "email": {
                    "title": "接收邮箱",
                    "type": "string",
                    "format": "email",
                    "description": "目标邮箱地址"
                  }
                },
                "required": [
                  "email"
                ]
              },
              "example": {
                "email": "garhlz257@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "发送成功响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 200
                    },
                    "msg": {
                      "title": "提示信息",
                      "type": "string",
                      "example": "验证码已发送"
                    },
                    "data": {
                      "title": "业务数据",
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 200,
                  "msg": "验证码已发送",
                  "data": {}
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 40302,
                  "message": "无权限访问",
                  "data": null
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 40401,
                  "message": "该账号未注册",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/reset-password": {
      "post": {
        "summary": "登录页重置密码",
        "deprecated": false,
        "description": "未登录状态下通过邮箱验证码重置密码。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "重置密码参数",
                "type": "object",
                "properties": {
                  "email": {
                    "title": "用户邮箱",
                    "type": "string"
                  },
                  "code": {
                    "title": "验证码",
                    "type": "string",
                    "description": "邮箱收到的验证码"
                  },
                  "newPassword": {
                    "title": "新密码",
                    "type": "string",
                    "description": "新设置的密码"
                  }
                },
                "required": [
                  "email",
                  "code",
                  "newPassword"
                ]
              },
              "example": {
                "email": "ffxlmf.klv40@yahoo.cn",
                "code": "61",
                "newPassword": "MrDkB4zZanUh7G0"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "重置成功响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 200
                    },
                    "msg": {
                      "title": "提示信息",
                      "type": "string",
                      "example": "重置成功"
                    },
                    "data": {
                      "title": "业务数据",
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 200,
                  "msg": "重置成功",
                  "data": {}
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "重置失败响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 400
                    },
                    "msg": {
                      "title": "错误提示",
                      "type": "string",
                      "example": "验证码错误或已过期"
                    },
                    "data": {
                      "title": "数据",
                      "type": "null"
                    }
                  }
                },
                "example": {
                  "code": 400,
                  "msg": "验证码错误或已过期",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "登出系统。",
        "tags": [
          "AdminAuth"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "退出成功响应",
                "type": "object",
                "properties": {
                  "code": {
                    "title": "状态码",
                    "type": "integer",
                    "example": 200
                  },
                  "msg": {
                    "title": "提示信息",
                    "type": "string",
                    "example": "已退出登录"
                  },
                  "data": {
                    "title": "业务数据",
                    "type": "object",
                    "properties": {},
                    "nullable": true
                  }
                },
                "required": [
                  "code",
                  "msg"
                ]
              },
              "example": {
                "code": 200,
                "msg": "已退出登录",
                "data": null
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "退出成功响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 200
                    },
                    "msg": {
                      "title": "提示信息",
                      "type": "string",
                      "example": "已退出登录"
                    },
                    "data": {
                      "title": "业务数据",
                      "type": "object",
                      "properties": {},
                      "nullable": true
                    }
                  },
                  "required": [
                    "code",
                    "msg"
                  ]
                },
                "example": {
                  "code": 200,
                  "msg": "已退出登录",
                  "data": {}
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "退出失败响应",
                  "type": "object",
                  "properties": {
                    "code": {
                      "title": "状态码",
                      "type": "integer",
                      "example": 401
                    },
                    "msg": {
                      "title": "错误提示",
                      "type": "string",
                      "example": "未登录或Token过期"
                    },
                    "data": {
                      "title": "数据",
                      "type": "null"
                    }
                  }
                },
                "example": {
                  "code": 401,
                  "msg": "未登录或Token过期",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/test/register": {
      "post": {
        "summary": "测试：创建管理员用户",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "nickname",
                  "password",
                  "email"
                ]
              },
              "example": {
                "nickname": "elaine",
                "password": "123456",
                "email": "garhlz257@163.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/auth/refresh": {
      "post": {
        "summary": "刷新token",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "refreshToken": "..."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isNewUser": {
                          "type": "null"
                        },
                        "hasPassword": {
                          "type": "null"
                        },
                        "userInfo": {
                          "type": "null"
                        },
                        "userSig": {
                          "type": "null"
                        },
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "isNewUser",
                        "hasPassword",
                        "userInfo",
                        "userSig",
                        "token",
                        "refreshToken"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 200,
                  "message": "操作成功",
                  "data": {
                    "isNewUser": null,
                    "hasPassword": null,
                    "userInfo": null,
                    "userSig": null,
                    "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE3LCJyb2xlIjoiVVNFUiIsIm5pY2tuYW1lIjoi5pig6K6w5rWL6K-V6LSm5Y-3Lea1i-ivleeJiCIsInR5cGUiOiJhY2Nlc3MiLCJleHAiOjE3NjY5MDk1NTR9.oHGva64ncj8Zw2v1YkonQzJNUJRokr5qnZ9JVYXYCX8",
                    "refreshToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjE3LCJ0eXBlIjoicmVmcmVzaCIsImV4cCI6MTc2ODg5Njc1NH0.vLRdH_bOujDVR41KLSASVPHJe7c8he0iIPRzLGAglQE"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
    "/api/admin/post/{postId}/audit-history": {
      "get": {
        "summary": "获取帖子审核历史",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "693c5c9951cd4f43d5167f5e",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/api/admin/post/{postId}": {
      "delete": {
        "summary": "管理员删除帖子",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "693c5c9951cd4f43d5167f5e",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "description": "",
            "required": false,
            "example": "测试删除帖子",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "jwt": []
    }
  ]
}